Redirected setup.py to use instances.py again, deleted instances.py and renamed instances_2.py to instances.py
