+++ aws_analysis_tools/cli/instances.py
@@ -1,179 +1,224 @@
-#!/usr/bin/env kaws-python
+# -*- coding: utf-8 -*-
+#
+# © 2016 Krux Digital, Inc.
+#
 
-import re
-import sys
-import logging
+#
+# Standard libraries
+#
 
-import boto.ec2
+from __future__ import absolute_import
 
-from texttable  import Texttable
-from pprint     import PrettyPrinter
-from optparse   import OptionParser
+#
+# Third party libraries
+#
 
-PP = PrettyPrinter( indent=2 )
+from texttable import Texttable
 
-###################
-### Arg parsing
-###################
+#
+# Internal libraries
+#
 
-parser = OptionParser("usage: %prog [options]" )
-parser.add_option(  "-v", "--verbose",      default=None, action="store_true",
-                    help="enable debug output" )
-parser.add_option(  "-H", "--no-header",    default=None, action="store_true",
-                    help="suppress table header" )
-parser.add_option(  "-r", "--region",       default='us-east-1',
-                    help="ec2 region to connect to" )
-parser.add_option(  "-g", "--group",        default=None,
-                    help="Include instances from these groups only (regex)" )
-parser.add_option(  "-G", "--exclude-group",default=None,
-                    help="Exclude instances from these groups (regex)" )
-parser.add_option(  "-n", "--name",         default=None,
-                    help="Include instances with these names only (regex)" )
-parser.add_option(  "-N", "--exclude-name", default=None,
-                    help="Exclude instances with these names (regex)" )
-parser.add_option(  "-t", "--type",         default=None,
-                    help="Include instances with these types only (regex)" )
-parser.add_option(  "-T", "--exclude-type", default=None,
-                    help="Exclude instances with these types (regex)" )
-parser.add_option(  "-z", "--zone",         default=None,
-                    help="Include instances with these zones only (regex)" )
-parser.add_option(  "-Z", "--exclude-zone", default=None,
-                    help="Exclude instances with these zones (regex)" )
-parser.add_option(  "-s", "--state",        default=None,
-                    help="Include instances with these states only (regex)" )
-parser.add_option(  "-S", "--exclude-state",default=None,
-                    help="Exclude instances with these states (regex)" )
+import krux_boto
+import krux_ec2.cli
+from krux_ec2.filter import Filter
+
+NAME = 'instances'
 
 
-(options, args) = parser.parse_args()
+class Application(krux_ec2.cli.Application):
 
-###################
-### Logging
-###################
+    # A dict with key=CLI options and value=AWS filters
+    _CLI_TO_AWS = {
+            'group': 'group-name',
+            'name': 'tag:Name',
+            'type': 'instance-type',
+            'zone': 'availability-zone',
+            'state': 'instance-state-name'
+        }
 
-if options.verbose: log_level = logging.DEBUG
-else:               log_level = logging.INFO
+    # List of all options
+    _OPTS = ['group', 'name', 'type', 'zone', 'state']
 
-logging.basicConfig(stream=sys.stdout, level=log_level)
-logging.basicConfig(stream=sys.stderr, level=(logging.ERROR,logging.CRITICAL))
+    # A dict with key=CLI options and value=instance attribute
+    _INSTANCE_ATTR = {
+            'group': lambda i, attr: next((g.name for g in i.groups if g.name == attr), None),
+            'name': lambda i, attr: i.tags.get('Name'),
+            'type': lambda i, attr: i.instance_type,
+            'zone': lambda i, attr: str(i._placement),
+            'state': lambda i, attr: i.state
+        }
 
+    def __init__(self, name=NAME):
+        # Call to the superclass to bootstrap.
+        super(Application, self).__init__(name=name)
 
-###################
-### Connection
-###################
+        # Change self.args into a dict so you can fetch values using
+        # keys instead of attributes
+        self.options = vars(self.args)
 
-conn = boto.ec2.connect_to_region( options.region )
+    def add_cli_arguments(self, parser):
+        # Call to the superclass first
+        super(Application, self).add_cli_arguments(parser)
 
-###################
-### Regexes
-###################
+        group = krux_ec2.cli.get_group(parser, self.name)
 
-regexes = {}
-for opt in [ 'group', 'exclude_group', 'name', 'exclude_name',
-             'type',  'exclude_type',  'zone', 'exclude_zone',
-             'state', 'exclude_state' ]:
+        group.add_argument(
+            "-H", "--no-header",
+            default=None,
+            action="store_true",
+            help="suppress table header",
+        )
 
-    ### we have a regex we should build
-    if options.__dict__.get( opt, None ):
-        regexes[ opt ] = re.compile( options.__dict__.get( opt ), re.IGNORECASE )
+        group.add_argument(
+            "-g", "--group",
+            default=None,
+            help="Include instances from these groups only (regex)",
+        )
 
-#PP.pprint( regexes )
+        group.add_argument(
+            "-G", "--exclude-group",
+            default=None,
+            help="Exclude instances from these groups (regex)",
+        )
 
-def get_instances():
-    instances   = [ i for r in conn.get_all_instances()
-                        for i in r.instances ]
+        group.add_argument(
+            "-n", "--name",
+            default=None,
+            help="Include instances with these names only (regex)",
+        )
 
-    rv          = [];
-    for i in instances:
+        group.add_argument(
+            "-N", "--exclude-name",
+            default=None,
+            help="Exclude instances with these names (regex)",
+        )
 
-        ### we will assume this node is one of the nodes we want
-        ### to operate on, and we will unset this flag if any of
-        ### the criteria fail
-        wanted_node = True
+        group.add_argument(
+            "-t", "--type",
+            default=None,
+            help="Include instances with these types only (regex)",
+        )
 
-        for re_name, regex in regexes.iteritems():
+        group.add_argument(
+            "-T", "--exclude-type",
+            default=None,
+            help="Exclude instances with these types (regex)",
+        )
 
-            ### What's the value we will be testing against?
-            if re.search( 'group', re_name ):
-                value = i.groups[0].name
-            elif re.search( 'name', re_name ):
-                value = i.tags.get( 'Name', '' )
-            elif re.search( 'type', re_name ):
-                value = i.instance_type
-            elif re.search( 'state', re_name ):
-                value = i.state
-            elif re.search( 'zone', re_name ):
-                ### i.region is an object. i._placement is a string.
-                value = str(i._placement)
+        group.add_argument(
+            "-z", "--zone",
+            default=None,
+            help="Include instances with these zones only (regex)",
+        )
 
+        group.add_argument(
+            "-Z", "--exclude-zone",
+            default=None,
+            help="Exclude instances with these zones (regex)",
+        )
+
+        group.add_argument(
+            "-s", "--state",
+            default=None,
+            help="Include instances with these states only (regex)",
+        )
+
+        group.add_argument(
+            "-S", "--exclude-state",
+            default=None,
+            help="Exclude instances with these states (regex)",
+        )
+
+    """
+    Convert options dictionary to use AWS filters as keys instead of CLI options.
+    """
+    def convert_args(self):
+        # Dictionary of options and values to put in the Filter
+        filter_dict = {}
+
+        # Add entries to filter_dict with key=AWS filters and value=option
+        # values for options that filter on inclusion
+        for opt in Application._OPTS:
+            if self.options[opt]:
+                aws_filter = Application._CLI_TO_AWS[opt]
+                filter_dict[aws_filter] = self.options[opt]
+
+        return filter_dict
+
+    """
+    Use filter_dict to filter instances based on inclusion/exclusion options
+    """
+    def filter_args(self, filter_dict):
+
+        f = Filter(filter_dict)
+
+        # Filter/find instances based on inclusion filters
+        instances = self.ec2.find_instances(f)
+
+        # Iterate through found instances and filter based on exclude options
+        for opt in Application._OPTS:
+            exclude_str = 'exclude_' + opt
+
+            if self.options[exclude_str]:
+                attribute = self.options[exclude_str]
             else:
-                logging.error( "Don't know what to do with: %s" % re_name )
                 continue
 
-            #PP.pprint( "name = %s value = %s pattern = %s" % ( re_name, value, regex.pattern ) )
+            # Exclude instances if they have an attribute that is excluded
+            instances = [
+                            i for i in instances if
+                            Application._INSTANCE_ATTR[opt](i, attribute) != attribute
+                        ]
 
-            ### Should the regex match or not match?
-            if re.search( 'exclude', re_name ):
-                rv_value = None
-            else:
-                rv_value = True
+        return instances
 
-            ### if the match is not what we expect, then clearly we
-            ### don't care about the node
-            result = regex.search( value )
+    """
+    Outputs filtered instances as a table
+    """
+    def output_table(self, instances):
+        table       = Texttable( max_width=0 )
 
-            ### we expected to get no results, excellent
-            if result == None and rv_value == None:
-                pass
+        table.set_deco( Texttable.HEADER )
+        table.set_cols_dtype( [ 't', 't', 't', 't', 't', 't', 't', 't' ] )
+        table.set_cols_align( [ 'l', 'l', 'l', 'l', 'l', 'l', 'l', 't' ] )
 
-            ### we expected to get some match, excellent
-            elif result is not None and rv_value is not None:
-                pass
+        if not self.args.no_header:
+            # using add_row, so the headers aren't being centered, for easier grepping
+            table.add_row(
+                [ '# id', 'Name', 'Type', 'Zone', 'Group', 'State', 'Root', 'Volumes' ] )
 
-            ### we don't care about this node
-            else:
-                wanted_node = False
-                break
+        for i in instances:
 
-        if wanted_node:
-            rv.append( i )
+            # XXX there's a bug where you can't get the size of the volumes, it's
+            # always reported as None :(
+            volumes = ", ".join( [ ebs.volume_id for ebs in i.block_device_mapping.values()
+                                    if ebs.delete_on_termination == False ] )
 
-    return rv
+            # you can use i.region instead of i._placement, but it pretty
+            # prints to RegionInfo:us-east-1. For now, use the private version
+            # XXX EVERY column in this output had better have a non-zero length
+            # or texttable blows up with 'width must be greater than 0' error
+            table.add_row( [ i.id, i.tags.get( 'Name', ' ' ), i.instance_type,
+                             i._placement , i.groups[0].name, i.state,
+                             i.root_device_type, volumes or '-' ] )
 
-def list_instances():
-    table       = Texttable( max_width=0 )
 
-    table.set_deco( Texttable.HEADER )
-    table.set_cols_dtype( [ 't', 't', 't', 't', 't', 't', 't', 't' ] )
-    table.set_cols_align( [ 'l', 'l', 'l', 'l', 'l', 'l', 'l', 't' ] )
+        # table.draw() blows up if there is nothing to print
+        if instances or not self.args.no_header:
+            print table.draw()
 
-    if not options.no_header:
-        ### using add_row, so the headers aren't being centered, for easier grepping
-        table.add_row(
-            [ '# id', 'Name', 'Type', 'Zone', 'Group', 'State', 'Root', 'Volumes' ] )
+    def run(self):
+        filter_dict = self.convert_args()
+        instances = self.filter_args(filter_dict)
+        self.output_table_to_file(instances)
 
-    instances = get_instances()
-    for i in instances:
 
-        ### XXX there's a bug where you can't get the size of the volumes, it's
-        ### always reported as None :(
-        volumes = ", ".join( [ ebs.volume_id for ebs in i.block_device_mapping.values()
-                                if ebs.delete_on_termination == False ] )
+def main():
+    app = Application()
+    # with app.context():
+    app.run()
 
-        ### you can use i.region instead of i._placement, but it pretty
-        ### prints to RegionInfo:us-east-1. For now, use the private version
-        ### XXX EVERY column in this output had better have a non-zero length
-        ### or texttable blows up with 'width must be greater than 0' error
-        table.add_row( [ i.id, i.tags.get( 'Name', ' ' ), i.instance_type,
-                         i._placement , i.groups[0].name, i.state,
-                         i.root_device_type, volumes or '-' ] )
-
-        #PP.pprint( i.__dict__ )
-
-    ### table.draw() blows up if there is nothing to print
-    if instances or not options.no_header:
-        print table.draw()
 
 if __name__ == '__main__':
-    list_instances()
-
+    main()
